requirements.txt
flask==3.0.0
gunicorn==21.2.0
gspread==5.12.0
google-auth==2.23.0
